Add files via upload
This is a landing page which is hand HTML/header CSS coded for simplicity. It holds a flexbox format view and some basic jav scripting, which activates the button on the page after a successful validation of the Google reCaptcha v2.0 checkbox. The Button redirects via an anchor link that includes a generated API key to the corresponding wordpress widget's passphrase, allowing access to my Information Technology career resume which is inaccessible when attempting direct browse from any other URL, this keeping information secure except for users who are attempting to browse to it from my main web site's home page.